Commit 5: Join lobby and show enemy player’s fighterType, username, and photo     - Add a way for an enemy to join or leave a lobby.     - Wrote lobby joiner’s data as enemy to the available lobby in database     - Refactored `GameLoadingViewModel` from using `@Observable` to conforming to `ObservableObject` instead in order to use Combine to implement waiting for an enemyPlayer     - Used `Combine` to load an enemyPlayer upon receiving a valid lobby     - Added a way to create an enemy player from a `GameLobby`

+++ iOS/FuFight-Old/FuFight/Game/GameLoading/GameLoadingViewModel.swift
@@ -7,25 +7,100 @@
 
 import SwiftUI
 import FirebaseAuth
+import Combine
+import FirebaseFirestore
 
-@Observable
-class GameLoadingViewModel: BaseAccountViewModel {
-    var player: Player
-    var enemyPlayer: Player?
-    var currentLobbyId: String?
+class GameLoadingViewModel: BaseAccountViewModel, ObservableObject {
+    @Published var lobby: GameLobby?
+    @Published var player: Player
+    @Published var enemyPlayer: Player?
+    @Published var currentLobbyId: String?
+    var onEnemyPlayerReceived: ((_ enemyPlayer: Player?)->Void)?
 
-    init(player: Player, enemyPlayer: Player? = nil, account: Account) {
+    private var isLobbyOwner: Bool = false
+    private var listener: ListenerRegistration?
+    private var cancellables = Set<AnyCancellable>()
+
+    init(player: Player, enemyPlayer: Player? = nil, account: Account, onEnemyPlayerReceived: ((_ enemyPlayer: Player?)->Void)? = nil) {
         self.player = player
         self.enemyPlayer = enemyPlayer
+        self.onEnemyPlayerReceived = onEnemyPlayerReceived
         super.init(account: account)
+
+        //After receiving a lobbyId (created or joined), subscribe to changes and update lobby
+        $currentLobbyId
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] lobbyId in
+                if lobbyId != nil {
+                    self?.unsubscribe()
+                    self?.subscribeToLobbyChanges()
+                }
+            }
+            .store(in: &cancellables)
+
+        //After getting a lobby with player and enemy, create an enemyPlayer
+        $lobby
+            .map { lobby in
+                Player(lobby: lobby, isLobbyOwner: self.isLobbyOwner)
+            }
+            .assign(to: \.enemyPlayer, on: self)
+            .store(in: &cancellables)
+
+        //After receiving creating an enemyPlayer, go to GameView
+        $enemyPlayer
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] player in
+                if player != nil {
+                    self?.onEnemyPlayerReceived?(player)
+                }
+            }
+            .store(in: &cancellables)
+    }
+
+    deinit {
+        unsubscribe()
+        lobby = nil
+        enemyPlayer = nil
+        currentLobbyId = nil
+    }
+
+    func unsubscribe() {
+        if listener != nil {
+            listener?.remove()
+            listener = nil
+        }
+    }
+
+    ///subscribe to lobby database changes and update VM's lobby based on changes
+    func subscribeToLobbyChanges() {
+        if listener != nil {
+            unsubscribe()
+        }
+        guard let currentLobbyId else { return }
+        let query = lobbiesDb.document(currentLobbyId)
+        listener = query
+            .addSnapshotListener { [weak self] snapshot, error in
+                guard let self = self else { return } //check if lobby creator goes here after enemy joins
+                if let snapshot, snapshot.exists {
+                    do {
+                        let fetchedLobby = try snapshot.data(as: GameLobby.self)
+                        if fetchedLobby.isValid {
+                            LOG("Created lobby is now valid against \(isLobbyOwner ? fetchedLobby.enemyUsername ?? "" : fetchedLobby.username) with fighter \(isLobbyOwner ? fetchedLobby.enemyFighterType?.rawValue ?? "" : fetchedLobby.fighterType?.rawValue ?? "")")
+                            lobby = fetchedLobby
+                        }
+                    } catch {
+                        LOGDE("Error creating lobby with error: \(error.localizedDescription)\t\t and data: \(snapshot.data()?.description ?? "")")
+                    }
+                }
+            }
     }
 
     //MARK: - ViewModel Overrides
 
     override func onAppear() {
         super.onAppear()
-        updateLoadingMessage(to: "Finding Opponent")
-        findOpponent()
+        updateLoadingMessage(to: "Finding opponent")
+        findOrCreateLobby()
     }
 
     override func onDisappear() {
@@ -34,31 +109,44 @@ class GameLoadingViewModel: BaseAccountViewModel {
     }
 
     //MARK: - Public Methods
-    func findOpponent() {
+    func findOrCreateLobby() {
         Task {
             do {
                 let lobbyIds = try await GameNetworkManager.findAvailableLobbies(userId: account.userId)
                 if lobbyIds.isEmpty {
-                    if let lobbyId = try await GameNetworkManager.createLobby(for: account) {
-                        currentLobbyId = lobbyId
-                        //TODO: Wait for up to 5-12 seconds
-                    }
+                    //Create lobby and wait for an enemy to join
+                    //Create an enemy player from kENEMYUSERNAME
+                    let lobbyId = try await GameNetworkManager.createLobby(for: account, fighterType: player.fighter.fighterType)
+                    isLobbyOwner = true
+                    updateLoadingMessage(to: "Waiting for opponent")
+                    currentLobbyId = lobbyId
+                    //TODO: Listen for changes and wait for up to 5-12 seconds
                 } else {
+                    //Join someone's lobby by writing to their lobby as an enemy
+                    //Create an enemy player from kUSERNAME
                     LOGD("Lobbies found at \(lobbyIds)")
+                    let lobbyId = lobbyIds.first!
+                    try await GameNetworkManager.joinLobby(lobbyId: lobbyId, account: account, fighterType: player.fighter.fighterType)
+                    isLobbyOwner = false
+                    updateLoadingMessage(to: "Syncing with opponent")
+                    currentLobbyId = lobbyId
                 }
             } catch {
                 updateError(MainError(type: .noOpponentFound, message: error.localizedDescription))
             }
         }
     }
-}
 
-//MARK: - Private Methods
-private extension GameLoadingViewModel {
     func deleteCurrentLobby() {
         guard let currentLobbyId else { return }
         Task {
-            try await GameNetworkManager.deleteCurrentLobby(lobbyId: currentLobbyId)
+            if let lobby, lobby.userId == player.userId {
+                //Only lobby's owner can delete the lobby
+                try await GameNetworkManager.deleteCurrentLobby(lobbyId: currentLobbyId)
+            } else {
+                //Delete enemy data from joined lobby
+                try await GameNetworkManager.leaveLobby(lobbyId: currentLobbyId)
+            }
         }
     }
 }
